+++ Examples/Event Sourcing/Examples.EventSourcing/Configuration/SwaggerConfig.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
 using Swashbuckle.AspNetCore.Swagger;
 
-// ReSharper disable once CheckNamespace
-namespace Microsoft.Extensions.DependencyInjection
+namespace Examples.Configuration
 {
     internal static class SwaggerConfig
     {